add property not found errors in controller

+++ controller/experiment.controller.js
@@ -1,5 +1,9 @@
 const {json} = require("express");
 const ExperimentRepository = require("../repositories/experiment.repository");
+const {PropertyNotFound} = require("../errors/NotFound.errors");
+const {InvalidProperty} = require("../errors/validation.errors");
+const {bodyValidator} = require("../validators/body.validator");
+const {ServerUnableError} = require("../errors/internal.errors");
 const experimentRepository = new ExperimentRepository();
 
 const getAllExperiments = async (req, res) => {
@@ -12,6 +16,8 @@ const getExperimentById = async (req, res) => {
         const experiment_id = req.params.experiment_id;
         const result = await experimentRepository.retrieve(experiment_id);
         res.status(200).json({result});
+    } else {
+        throw new PropertyNotFound("experiment_id");
     }
 }
 
@@ -20,26 +26,40 @@ const getExperimentsByAccountId = async (req, res) => {
         const account_id = req.params.account_id;
         const experiments = await experimentRepository.findByAccount("account_id", account_id);
         res.status(200).json({experiments});
+    } else {
+        throw new PropertyNotFound("account_id");
     }
 }
 
 const getExperimentsAB = async (req, res) => {
-    const account_id = req.params.account_id;
-    const result = await experimentRepository.findGroup("type", "a-b", "account_id", account_id);
-    res.status(200).json({result});
+    if (req.params.account_id) {
+        const account_id = req.params.account_id;
+        const result = await experimentRepository.findGroup("type", "a-b", "account_id", account_id);
+        res.status(200).json({result});
+    } else {
+        throw new PropertyNotFound("account_id");
+    }
 }
 
 const getExperimentsFF = async (req, res) => {
-    const account_id = req.params.account_id;
-    const result = await experimentRepository.findGroup("type", "f-f", "account_id", account_id);
-    res.status(200).json({result});
+    if (req.params.account_id) {
+        const account_id = req.params.account_id;
+        const result = await experimentRepository.findGroup("type", "f-f", "account_id", account_id);
+        res.status(200).json({result});
+    } else {
+        throw new PropertyNotFound("account_id");
+    }
 }
 
 const getExperimentsByDate = async (req, res) => {
-    const year = req.query.year;
-    const month = req.query.month;
-    const result = await experimentRepository.findByDate(year, month);
-    res.status(200).json({result});
+    if (req.query.year && req.query.month) {
+        const year = req.query.year;
+        const month = req.query.month;
+        const result = await experimentRepository.findByDate(year, month);
+        res.status(200).json({result});
+    } else {
+        throw new PropertyNotFound("year and month");
+    }
 }
 
 const createExperiments = async (req, res) => {
@@ -48,17 +68,24 @@ const createExperiments = async (req, res) => {
 }
 
 const updateExperimentsByID = async (req, res) => {
-    const experimentID = req.params.experiment_id;
-    const result = await experimentRepository.update(experimentID, req.body)
-    res.status(200).json({result});
+    if (req.params.experiment_id) {
+        const experimentID = req.params.experiment_id;
+        const result = await experimentRepository.update(experimentID, req.body)
+        res.status(200).json({result});
+    } else {
+        throw new PropertyNotFound("experiment_id");
+    }
 }
 
 const deleteExperimentsByID = async (req, res) => {
-    const experimentID = req.params.experiment_id;
-    const result = await experimentRepository.delete(experimentID)
-    res.status(200).json({result});
+    if (req.params.experiment_id) {
+        const experimentID = req.params.experiment_id;
+        const result = await experimentRepository.delete(experimentID)
+        res.status(200).json({result});
+    } else {
+        throw new PropertyNotFound("experiment_id");
+    }
 }
-
 
 
 module.exports = {
@@ -66,7 +93,7 @@ module.exports = {
     getExperimentById,
     getExperimentsByAccountId,
     getExperimentsAB,
-    getExperimentsFF,,
+    getExperimentsFF,
     getExperimentsByDate,
     updateExperimentsByID,
     deleteExperimentsByID,
